backend/requirements.txt
asgiref==3.9.1
diff-match-patch==20241021
Django==5.2.4
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.8
django-js-asset==3.1.2
django-redis==6.0.0
django-unfold==0.62.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
pillow==11.3.0
psycopg2-binary==2.9.10
redis==6.2.0
sqlparse==0.5.3
tablib==3.8.0
